requirements.txt
discord.py==2.4.0
anthropic==0.40.0
APScheduler==3.10.4
python-dotenv==1.0.1